requirements.txt
flask
Flask-PyMongo
Flask
flask_restful
python-dotenv
pymongo[srv]
schema
flask_cors